Crate wallet_service_mirror_private

The private side of wallet-service-mirror. This program forms outgoing connections to both a wallet service instance, as well as a public wallet-service-mirror instance. It then proceeds to poll the public side of the mirror for requests which it then forwards to the wallet service. When a response is received it is then forwarded back to the mirror.

Modules

crypto ğŸ”’

Cryptographic primitives.

Structs

Config

Command line config

MonitorId

A wrapper to ease monitor id parsing from a hex string when using StructOpt.

Constants

FULL_SERVICE_TIMEOUT ğŸ”’

How long do we wait for full-service to reply?

SUPPORTED_ENDPOINTS ğŸ”’

Statics

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

LOC ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

RS ğŸ”’

Functions

main ğŸ”’

parse_duration_in_milliseconds ğŸ”’

process_encrypted_request ğŸ”’

process_unencrypted_request ğŸ”’

validate_method ğŸ”’